+++ Asset.API/Controllers/WorkOrderStatusController.cs
@@ -1,14 +1,15 @@
 ﻿using Asset.API.Helpers;
 using Asset.Domain.Services;
+using Asset.Models;
 using Asset.ViewModels.WorkOrderStatusVM;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace Asset.API.Controllers
 {
@@ -16,31 +17,32 @@ namespace Asset.API.Controllers
     [ApiController]
     public class WorkOrderStatusController : ControllerBase
     {
+        private IWorkOrderTrackingService _workOrderTrackingService;
         private IWorkOrderStatusService _workOrderStatusService;
 
-        public WorkOrderStatusController(IWorkOrderStatusService workOrderStatusService)
+        public WorkOrderStatusController(IWorkOrderStatusService workOrderStatusService, IWorkOrderTrackingService workOrderTrackingService)
         {
+            _workOrderTrackingService = workOrderTrackingService;
             _workOrderStatusService = workOrderStatusService;
         }
-        // GET: api/<WorkOrderStatusController>
+
         [HttpGet]
         public IEnumerable<IndexWorkOrderStatusVM> Get()
         {
             return _workOrderStatusService.GetAllWorkOrderStatuses();
         }
 
-        // GET api/<WorkOrderStatusController>/5
         [HttpGet("{id}")]
         public IndexWorkOrderStatusVM Get(int id)
         {
             return _workOrderStatusService.GetWorkOrderStatusById(id);
         }
 
-        // POST api/<WorkOrderStatusController>
+    
         [HttpPost]
         public IActionResult Post(CreateWorkOrderStatusVM createWorkOrderStatusVM)
         {
-            var lstcodes = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.Name == createWorkOrderStatusVM.Name).ToList();
+            var lstcodes = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.Code == createWorkOrderStatusVM.Code).ToList();
             if (lstcodes.Count > 0)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError, new Response { Status = "name", Message = " name already exist", MessageAr = "هذا الاسم مسجل سابقاً" });
@@ -64,19 +66,21 @@ namespace Asset.API.Controllers
 
         // PUT api/<WorkOrderStatusController>/5
         [HttpPut("{id}")]
+        [Route("UpdateWorkOrderStatus")]
         public IActionResult Put(int id, EditWorkOrderStatusVM editWorkOrderStatusVM)
         {
-            var lstcodes = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.Code == editWorkOrderStatusVM.Code && a.Id == editWorkOrderStatusVM.Id).ToList();
+            
+            var lstcodes = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.Code == editWorkOrderStatusVM.Code && a.Id != editWorkOrderStatusVM.Id).ToList();
             if (lstcodes.Count > 0)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError, new Response { Status = "code", Message = " name already exist", MessageAr = "هذا الاسم مسجل سابقاً" });
             }
-            var lstNames = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.Name == editWorkOrderStatusVM.Name && a.Id == editWorkOrderStatusVM.Id).ToList();
+            var lstNames = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.Name == editWorkOrderStatusVM.Name && a.Id != editWorkOrderStatusVM.Id).ToList();
             if (lstNames.Count > 0)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError, new Response { Status = "name", Message = " name already exist", MessageAr = "هذا الاسم مسجل سابقاً" });
             }
-            var lstArNames = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.NameAr == editWorkOrderStatusVM.NameAr && a.Id == editWorkOrderStatusVM.Id).ToList();
+            var lstArNames = _workOrderStatusService.GetAllWorkOrderStatuses().ToList().Where(a => a.NameAr == editWorkOrderStatusVM.NameAr && a.Id != editWorkOrderStatusVM.Id).ToList();
             if (lstArNames.Count > 0)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError, new Response { Status = "nameAr", Message = " arabic name already exist", MessageAr = "هذا الاسم مسجل سابقاً" });
@@ -88,11 +92,29 @@ namespace Asset.API.Controllers
             }
         }
 
-        // DELETE api/<WorkOrderStatusController>/5
+    
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public ActionResult<WorkOrderStatus> Delete(int id)
         {
-            _workOrderStatusService.DeleteWorkOrderStatus(id);
+            try
+            {
+                var lstStatus = _workOrderTrackingService.GetAll().ToList().Where(a => a.WorkOrderStatusId == id).ToList();
+                if (lstStatus.Count > 0)
+                {
+                    return StatusCode(StatusCodes.Status500InternalServerError, new Response { Status = "wostatus", Message = "This status already in use", MessageAr = "هذه الحالة مستخدمة" });
+                }
+                else
+                {
+                    _workOrderStatusService.DeleteWorkOrderStatus(id);
+                }
+            }
+            catch (DbUpdateConcurrencyException ex)
+            {
+                string msg = ex.Message;
+                return BadRequest("Error in delete");
+            }
+
+            return Ok();
         }
     }
 }
